+++ engine/launcher/src/main/scala/org/enso/launcher/distribution/DefaultManagers.scala
@@ -1,5 +1,10 @@
 package org.enso.launcher.distribution
 
+import org.enso.distribution.PortableDistributionManager
+import org.enso.distribution.locking.{
+  ResourceManager,
+  ThreadSafeFileLockManager
+}
 import org.enso.launcher.cli.{
   CLIRuntimeVersionManagementUserInterface,
   GlobalCLIOptions
@@ -11,14 +16,7 @@ import org.enso.runtimeversionmanager.components.{
   RuntimeComponentUpdaterFactory,
   RuntimeVersionManager
 }
-import org.enso.runtimeversionmanager.distribution.{
-  PortableDistributionManager,
-  TemporaryDirectoryManager
-}
-import org.enso.runtimeversionmanager.locking.{
-  ResourceManager,
-  ThreadSafeFileLockManager
-}
+import org.enso.runtimeversionmanager.distribution.TemporaryDirectoryManager
 import org.enso.runtimeversionmanager.releases.engine.EngineRepository
 import org.enso.runtimeversionmanager.releases.graalvm.GraalCEReleaseProvider