+++ engine/launcher/src/main/scala/org/enso/launcher/releases/launcher/LauncherManifest.scala
@@ -3,7 +3,7 @@ package org.enso.launcher.releases.launcher
 import io.circe.{yaml, Decoder}
 import nl.gn0s1s.bump.SemVer
 import org.enso.runtimeversionmanager.releases.ReleaseProviderException
-import org.enso.pkg.SemVerJson._
+import org.enso.editions.SemVerJson._
 
 import scala.util.{Failure, Try}