+++ engine/launcher/src/test/scala/org/enso/launcher/releases/fallback/FileStorageFallbackReleaseProviderSpec.scala
@@ -1,8 +1,8 @@
 package org.enso.launcher.releases.fallback
 
-import java.nio.file.Path
+import org.enso.distribution.FileSystem
 
-import org.enso.runtimeversionmanager.FileSystem
+import java.nio.file.Path
 import org.enso.launcher.TestHelpers
 import org.enso.launcher.releases.fallback.staticwebsite.FileStorageFallbackReleaseProvider
 import org.scalatest.matchers.should.Matchers