+++ lib/scala/project-manager/src/main/scala/org/enso/projectmanager/versionmanagement/DistributionConfiguration.scala
@@ -1,15 +1,12 @@
 package org.enso.projectmanager.versionmanagement
 
-import org.enso.runtimeversionmanager.Environment
+import org.enso.distribution.locking.ResourceManager
+import org.enso.distribution.{DistributionManager, EditionManager, Environment}
 import org.enso.runtimeversionmanager.components.{
   RuntimeVersionManagementUserInterface,
   RuntimeVersionManager
 }
-import org.enso.runtimeversionmanager.distribution.{
-  DistributionManager,
-  TemporaryDirectoryManager
-}
-import org.enso.runtimeversionmanager.locking.ResourceManager
+import org.enso.runtimeversionmanager.distribution.TemporaryDirectoryManager
 import org.enso.runtimeversionmanager.releases.ReleaseProvider
 import org.enso.runtimeversionmanager.releases.engine.EngineRelease
 import org.enso.runtimeversionmanager.runner.JVMSettings
@@ -30,6 +27,9 @@ trait DistributionConfiguration {
 
   /** A [[ResourceManager]] instance. */
   def resourceManager: ResourceManager
+
+  /** An [[EditionManager]] instance. */
+  def editionManager: EditionManager
 
   /** A [[TemporaryDirectoryManager]] instance. */
   def temporaryDirectoryManager: TemporaryDirectoryManager